Add components for the header and map
I recreated the header as a Svelte component and partially converted the
HrrrMap web component to a svelte component. I need to get the lifecycle
callbacks properly working so that it will load and render the map
