+++ services/frontend/src/App.svelte
@@ -1,11 +1,7 @@
 <script>
-	let count = 0;
-
-	function handleClick() {
-		count += 1;
-	}
+  import Header from "./components/Header.svelte";
+  import HrrrMap from "./components/HrrrMap.svelte";
 </script>
-<h1>Hello, world</h1>
-<button on:click={handleClick}>
-	Clicked {count} {count === 1 ? 'time' : 'times'}
-</button>
+
+<Header />
+<HrrrMap />
