+++ services/frontend/src/components/Header.svelte
@@ -1,0 +1,7 @@
+<header class="cluster">
+  <h1>
+    <!-- FIXME: This should use a utiilty class to set display flex -->
+    <a href="/" style="display:flex"><b>HRRR</b> Smoke</a>
+  </h1>
+  <p><small>Experimental</small></p>
+</header>
